Functioning Fortran interface.

+++ src/c++/hashtable.cpp
@@ -39,8 +39,7 @@ HashTable::HashTable(int const tid)
 
 size_t HashTable::query_insert(std::string_view region_name) noexcept
 {
-//TODO - Needs a different hashing algorithm that is compatible with Fortran. this used to return size_t
-  int hash = hash_function_(region_name);
+  size_t hash = hash_function_(region_name);
 
   if (table_.count(hash) == 0){
     table_.emplace(hash,HashEntry(region_name));
